+++ crates/model/src/orderbook/own.rs
@@ -909,15 +909,15 @@ impl OwnBookLevel {
         self.orders.insert(order.client_order_id, order);
     }
 
-    /// Updates an existing order at this price level. Updated order must match the level's price.
-    /// Removes the order if size becomes zero.
+    /// Updates an order at this price level, inserting it if missing. Updated order
+    /// must match the level's price. Removes the order if the size becomes zero.
     pub fn update(&mut self, order: OwnBookOrder) {
         debug_assert_eq!(order.price, self.price.value);
 
         if order.size.is_zero() {
             self.orders.shift_remove(&order.client_order_id);
         } else {
-            self.orders[&order.client_order_id] = order;
+            self.orders.insert(order.client_order_id, order);
         }
     }